# lib/api-zod/src/generated/types/artStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MintRecord } from "./mintRecord";

export interface ArtStats {
  totalMints: number;
  totalDeployments: number;
  uniqueArtists: number;
  recentMints: MintRecord[];
}