# lib/api-zod/src/generated/types/deployment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Deployment {
  id: number;
  contractAddress: string;
  chainId: number;
  network: string;
  deployerAddress: string;
  mintPrice: string;
  createdAt: Date;
}