# lib/api-zod/src/generated/types/mintRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MintRecord {
  id: number;
  deploymentId: number;
  tokenId: number;
  title: string;
  artistAddress: string;
  txHash: string;
  mintedAt: Date;
}