# lib/api-zod/src/generated/types/recordMintBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecordMintBody {
  deploymentId: number;
  tokenId: number;
  title: string;
  artistAddress: string;
  txHash: string;
}